+++ test/e2e/baseline_test.go
@@ -38,12 +38,8 @@ import (
 	"istio.io/istio/pkg/test/framework/components/echo"
 	"istio.io/istio/pkg/test/framework/components/echo/check"
 	"istio.io/istio/pkg/test/framework/components/echo/common/ports"
-	"istio.io/istio/pkg/test/framework/components/echo/util/traffic"
-	kubetest "istio.io/istio/pkg/test/kube"
 	"istio.io/istio/pkg/test/scopes"
-	"istio.io/istio/pkg/test/util/retry"
 	"istio.io/istio/pkg/util/sets"
-	appsv1 "k8s.io/api/apps/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 )
@@ -576,73 +572,6 @@ func TestRemoveAddNsOrServiceWaypoint(t *testing.T) {
 			})
 		})
 	}
-}
-
-func TestKmeshRestart(t *testing.T) {
-	framework.NewTest(t).Run(func(t framework.TestContext) {
-		src := apps.EnrolledToKmesh[0]
-		dst := apps.ServiceWithWaypointAtServiceGranularity
-		options := echo.CallOptions{
-			To:    dst,
-			Count: 1,
-			// Determine whether it is managed by Kmesh by passing through Waypoint.
-			Check: httpValidator,
-			Port: echo.Port{
-				Name: "http",
-			},
-			Retry: echo.Retry{NoRetry: true},
-		}
-
-		g := traffic.NewGenerator(t, traffic.Config{
-			Source:   src,
-			Options:  options,
-			Interval: 50 * time.Millisecond,
-		}).Start()
-
-		restartKmesh(t)
-
-		g.Stop().CheckSuccessRate(t, 1)
-	})
-}
-
-func restartKmesh(t framework.TestContext) {
-	patchOpts := metav1.PatchOptions{}
-	patchData := fmt.Sprintf(`{
-			"spec": {
-				"template": {
-					"metadata": {
-						"annotations": {
-							"kubectl.kubernetes.io/restartedAt": %q
-						}
-					}
-				}
-			}
-		}`, time.Now().Format(time.RFC3339))
-	ds := t.Clusters().Default().Kube().AppsV1().DaemonSets(KmeshNamespace)
-	_, err := ds.Patch(context.Background(), KmeshDaemonsetName, types.StrategicMergePatchType, []byte(patchData), patchOpts)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if err := retry.UntilSuccess(func() error {
-		d, err := ds.Get(context.Background(), KmeshDaemonsetName, metav1.GetOptions{})
-		if err != nil {
-			return err
-		}
-		if !daemonsetsetComplete(d) {
-			return fmt.Errorf("rollout is not yet done")
-		}
-		return nil
-	}, retry.Timeout(60*time.Second), retry.Delay(2*time.Second)); err != nil {
-		t.Fatal("failed to wait for Kmesh rollout status for: %v", err)
-	}
-	if _, err := kubetest.CheckPodsAreReady(kubetest.NewPodFetch(t.AllClusters()[0], KmeshNamespace, "app=kmesh")); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func daemonsetsetComplete(ds *appsv1.DaemonSet) bool {
-	return ds.Status.UpdatedNumberScheduled == ds.Status.DesiredNumberScheduled && ds.Status.NumberReady == ds.Status.DesiredNumberScheduled && ds.Status.ObservedGeneration >= ds.Generation
 }
 
 func runTest(t *testing.T, f func(t framework.TestContext, src echo.Instance, dst echo.Instance, opt echo.CallOptions)) {
